Ninth Commit
I showed the number of products that users added to their shopping cart.Users can see the number of products in their shopping cart from the menu.

I created a class named ShoppingCartViewModel.In this class, I used both shopping carts and orders in a single class.

I created a controller named ShoppingCartController.

I created Index View for ShoppingCartController and I edited the Index View.

I performed the operation of increasing the products in the shopping cart, decreasing the products in the shopping cart and deleting the products in the shopping cart.

I performed the email confirmation process on the cart screen for users who placed an order without confirming their email.